unit tests for utf8 validation
