+++ src/llm/http_llm_calculator.cc
@@ -32,6 +32,7 @@
 #include <rapidjson/writer.h>
 
 #include "../profiler.hpp"
+#include "../stringutils.hpp"
 #include "http_payload.hpp"
 #include "llmnoderesources.hpp"
 
@@ -358,37 +359,6 @@ class TextStreamer {
     std::vector<int64_t> tokenCache;
     size_t printLen{0};
 
-private:
-    bool isValidUtf8(const std::string& text) {
-        // inspect the chars from the end of the string to test if the utf8 sequence is complete
-        int byte_counter = 0;
-        for (int i = text.size() - 1; i >= 0 && byte_counter <= 3; i--) {
-            int x = static_cast<int>(static_cast<unsigned char>(text[i]));
-            if ((x >> 7) == 0b0)
-                return true;         // last char is a single byte char
-            if ((x >> 6) == 0b10) {  // octet belong to multibyte sequence
-                byte_counter++;
-            } else if ((x >> 5) == 0b110) {  // first byte of 2 byte sequence
-                if (byte_counter + 1 == 2)
-                    return true;
-                else
-                    return false;
-            } else if ((x >> 4) == 0b1110) {  // first byte of 3 byte sequence
-                if (byte_counter + 1 == 3)
-                    return true;
-                else
-                    return false;
-            } else if ((x >> 3) == 0b11110) {  // first byte of 3 byte sequence
-                if (byte_counter + 1 == 4)
-                    return true;
-                else
-                    return false;
-            } else
-                return false;
-        }
-        return true;
-    }
-
 public:
     TextStreamer(std::shared_ptr<Tokenizer> tokenizer) :
         tokenizer(tokenizer) {}
@@ -420,7 +390,7 @@ public:
         }
         return std::nullopt;
     }
-}
+};
 
 static bool
 applyChatTemplate(TextProcessor& textProcessor, std::string modelsPath, std::string& requestBody, std::string& output) {
